move getSecondsUntilMidnightOrMinimum
