+++ src/place-screen/utils.ts
@@ -1,5 +1,5 @@
 import {StopPlacesMode} from '../nearby-stop-places';
-import {getSecondsUntilMidnightOrMinimum} from './hooks/use-departures-data';
+import {addDays, differenceInSeconds, parseISO} from 'date-fns';
 
 const MIN_TIME_RANGE = 3 * 60 * 60; // Three hours
 const ONE_WEEK_TIME_RANGE = 7 * 24 * 60 * 60;
@@ -22,3 +22,17 @@ export const getTimeRangeByMode = (mode: StopPlacesMode, startTime?: string) =>
         startTime ?? new Date().toISOString(),
         MIN_TIME_RANGE,
       );
+
+/**
+ * Get seconds until midnight, but a minimum of `minSeconds`
+ */
+function getSecondsUntilMidnightOrMinimum(
+  isoTime: string,
+  minimumSeconds: number = 0,
+): number {
+  const timeUntilMidnight = differenceInSeconds(
+    addDays(parseISO(isoTime), 1).setHours(0, 0, 0),
+    parseISO(isoTime),
+  );
+  return Math.round(Math.max(timeUntilMidnight, minimumSeconds));
+}
